Repairing format specifier - variable type mismatches
git-svn-id: https://stealth.svn.sourceforge.net/svnroot/yodl/trunk@121 7de777c3-3b1c-0410-83c5-80958bd9f4b1

+++ yodl/src/hashmap/hashmapkeys.c
@@ -4,7 +4,7 @@ void hashmap_keys(register HashMap *map)
 {
     size_t idx;
 
-    fprintf(stderr, "There are %u active keys\n", map->d_n);
+    fprintf(stderr, "There are %u active keys\n", (unsigned)map->d_n);
 
     for (idx = 0; idx < map->d_size; idx++)
     {
@@ -16,11 +16,11 @@ void hashmap_keys(register HashMap *map)
             continue;
 
             case REMOVED:
-                fprintf(stderr, "   %3d: <removed>\n", idx);
+                fprintf(stderr, "   %3u: <removed>\n", (unsigned)idx);
             break;
 
             default:
-                fprintf(stderr, "   %3d: `%s'\n", idx, atIdx->d_key);
+                fprintf(stderr, "   %3u: `%s'\n", (unsigned)idx, atIdx->d_key);
             break;
         }
     }
