+++ yodl/src/yodlpost/handlexmltocentry.c
@@ -4,5 +4,5 @@ void handle_xml_tocentry(long offset, HashItem *item)
 {
     file_copy2offset(global.d_out, postqueue_istream(), offset);
     fprintf(global.d_out, "<a id=\"l%u\"></a>\n",
-                                        (size_t)hashitem_value(item));
+                                        (unsigned)hashitem_value(item));
 }
